+++ packages/appcd-config/test/test-config.js
@@ -1,6 +1,16 @@
 import AppcdConfig from '../dist/config';
 
 describe('AppcdConfig', () => {
+	it('should error if options is invalid', () => {
+		expect(() => {
+			new AppcdConfig(null);
+		}).to.throw(TypeError, 'Expected options to be an object');
+
+		expect(() => {
+			new AppcdConfig(123);
+		}).to.throw(TypeError, 'Expected options to be an object');
+	});
+
 	it('should create a new config instance', () => {
 		const cfg = new AppcdConfig();
 		cfg.set('foo', 'bar');