+++ packages/appcd-core/CHANGELOG.md
@@ -2,7 +2,7 @@
 
  * BREAKING CHANGE: Requires Node.js 10.13.0 or newer.
    [(DAEMON-334)](https://jira.appcelerator.org/browse/DAEMON-334)
- * BREAKING CHANGE: Bumped core Node.js version to 14.13.0. This affects all plugins.
+ * BREAKING CHANGE: Bumped core Node.js version to 14.15.0 LTS. This affects all plugins.
    [(DAEMON-319)](https://jira.appcelerator.org/browse/DAEMON-319)
  * BREAKING CHANGE: Main file exports library instead of the main entry point.
  * BREAKING CHANGE(status-monitor): Daemon uptime has changed from seconds to milliseconds.
@@ -32,6 +32,7 @@
  * fix(telemetry): Lowered telemetry send timeout from 1 minute to 10 seconds to prevent a long
    hang during telemetry shutdown while it waits to send a batch of events.
  * fix(server): Decouple shutdown trigger from server by moving to main entry script.
+ * fix(server): Replaced AMPLIFY CLI references with Axway CLI.
  * fix(plugin): Added HTTPS proxy check and error if strictSSL is disabled.
  * chore: Updated dependencies.