+++ packages/appcd/CHANGELOG.md
@@ -17,6 +17,7 @@
  * feat: Added additional documentation and examples to the help output for some commands.
  * feat: Show aliases in help for commands such as `config` and `pm`.
  * feat(status): Added new "Health" section to status output.
+ * feat: Added notificaiton if new version is available.
  * refactor: Updated to latest `AppcdConfig` usage where `save()` is now synchronous and we no
    longer need to set the config file to save to since it uses the same file path that was used to
    load the config.